LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/datetime-weekmonth-expected.txt

Date and Time Inputs


PASS month type support on input element
PASS [month] The value must be a value that is a valid global date and time string
PASS [month] The min attribute must have a value that is a valid global date and time string
PASS [month] The max attribute must have a value that is a valid global date and time string
PASS [month] The step attribute must be expressed in seconds
PASS [month] stepUp method support on input 'month' element
PASS [month] stepDown method support on input 'month' element
PASS week type support on input element
PASS [week] The value must be a value that is a valid global date and time string
PASS [week] The min attribute must have a value that is a valid global date and time string
PASS [week] The max attribute must have a value that is a valid global date and time string
PASS [week] The step attribute must be expressed in seconds
PASS [week] stepUp method support on input 'week' element
PASS [week] stepDown method support on input 'week' element
